requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-filter==23.1
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
Pillow==10.0.0
python-decouple==3.8
